fix voice cache population

+++ src/main/java/union/utils/database/managers/VoiceChannelManager.java
@@ -1,5 +1,9 @@
 package union.utils.database.managers;
 
+import com.github.benmanes.caffeine.cache.Cache;
+import com.github.benmanes.caffeine.cache.Caffeine;
+import net.dv8tion.jda.api.JDA;
+import net.dv8tion.jda.api.entities.channel.concrete.VoiceChannel;
 import union.utils.database.ConnectionUtil;
 import union.utils.database.LiteDBBase;
 
@@ -9,14 +13,18 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.stream.Collectors;
+
+import static union.utils.CastUtil.castLong;
 
 public class VoiceChannelManager extends LiteDBBase {
 	// Cache
-	private final HashMap<Long, Long> cache = new HashMap<>();
+	// user - channel
+	private final Cache<Long, Long> cache = Caffeine.newBuilder()
+		.build();
 
 	public VoiceChannelManager(ConnectionUtil cu) {
 		super(cu, "voiceChannels");
+		// try to populate cache
 		cache.putAll(getDbCache());
 	}
 
@@ -28,46 +36,70 @@ public class VoiceChannelManager extends LiteDBBase {
 	}
 
 	public void remove(long channelId) {
-		Optional.ofNullable(getUser(channelId)).ifPresent(cache::remove);
-		try {
-			execute("DELETE FROM %s WHERE (channelId=%d)".formatted(table, channelId));
-		} catch (SQLException ignored) {}
+		Optional.ofNullable(getUser(channelId))
+			.ifPresentOrElse(userId -> {
+				cache.invalidate(userId);
+				try {
+					execute("DELETE FROM %s WHERE (userId=%d OR channelId=%d)".formatted(table, userId, channelId));
+				} catch (SQLException ignored) {}
+			}, () -> {
+				try {
+					execute("DELETE FROM %s WHERE (channelId=%d)".formatted(table, channelId));
+				} catch (SQLException ignored) {}
+			});
 	}
 
 	public boolean existsUser(long userId) {
-		return cache.containsKey(userId);
+		return cache.getIfPresent(userId) != null;
 	}
 
 	public boolean existsChannel(long channelId) {
-		return cache.containsValue(channelId);
+		return cache.getIfPresent(channelId) != null;
 	}
 
 	public void setUser(long userId, long channelId) throws SQLException {
 		// Remove user with same channelId
-		cache.entrySet().stream()
-			.filter((e) -> e.getValue().equals(channelId))
-			.map(Map.Entry::getKey)
-			.findFirst().ifPresent(cache::remove);
+		Long cacheChannelId = cache.getIfPresent(userId);
+		if (cacheChannelId != null && channelId == cacheChannelId) {
+			cache.invalidate(userId);
+		}
 		// Add new user
 		cache.put(userId, channelId);
 		execute("UPDATE %s SET userId=%s WHERE (channelId=%d)".formatted(table, userId, channelId));
 	}
 
 	public Long getChannel(long userId) {
-		return cache.get(userId);
+		return cache.getIfPresent(userId);
 	}
 
 	public Long getUser(long channelId) {
-		return cache.entrySet().stream()
-			.filter((e) -> e.getValue().equals(channelId))
+		return cache.asMap().entrySet()
+			.stream()
+			.filter(e -> e.getValue() == channelId)
+			.findFirst()
 			.map(Map.Entry::getKey)
-			.findAny().orElse(null);
+			.orElse(null);
 	}
 
 	private Map<Long, Long> getDbCache() {
 		List<Map<String, Object>> data = select("SELECT * FROM %s".formatted(table), Set.of("channelId", "userId"));
 		if (data.isEmpty()) return Map.of();
-		return data.stream()
-			.collect(Collectors.toMap(s -> (Long) s.get("userId"), s -> (Long) s.get("channelId")));
+		Map<Long, Long> cacheData = new HashMap<>();
+		for (Map<String, Object> row : data) {
+			cacheData.put(castLong(row.get("userId")), castLong(row.get("channelId")));
+		}
+		return cacheData;
+	}
+
+	public void checkCache(JDA jda) {
+		cache.asMap().forEach((userId, channelId) -> {
+			VoiceChannel voiceChannel = jda.getVoiceChannelById(channelId);
+			if (voiceChannel == null) {
+				remove(channelId);
+			} else if (voiceChannel.getMembers().isEmpty()) {
+				voiceChannel.delete().queue();
+				remove(channelId);
+			}
+		});
 	}
 }
